Renamed Split public API components to follow Swift guidelines. Added Factory builder component. Added Factory builder tests. Fixed Swift example to work with new API definition
