+++ Example/Split/GetTreatmentViewController.swift
@@ -21,8 +21,8 @@ class GetTreatmentViewController: UIViewController {
     @IBOutlet weak var evaluateActivityIndicator: UIActivityIndicatorView!
     
     var factory: SplitFactory?
-    var client: SplitClientProtocol?
-    var manager: SplitManagerProtocol?
+    var client: SplitClient?
+    var manager: SplitManager?
 
     @IBAction func evaluate(_ sender: Any) {
         evaluate()
@@ -71,16 +71,17 @@ class GetTreatmentViewController: UIViewController {
         let key: Key = Key(matchingKey: matchingKeyText, bucketingKey: bucketing)
         
         //Split Factory
-        self.factory = SplitFactory(apiKey: authorizationKey, key: key, config: config)
+        let builder = DefaultSplitFactoryBuilder()
+        self.factory = builder.setApiKey(authorizationKey).setKey(key).setConfig(config).build()
         
         //Split Client
-        self.client = self.factory?.client()
+        self.client = self.factory?.client
         
         //Split Manager
-        self.manager = self.factory?.manager()
+        self.manager = self.factory?.manager
         
         //Showing sdk version in UI
-        self.sdkVersion?.text = "SDK Version: \(self.factory?.version() ?? "unknown") "
+        self.sdkVersion?.text = "SDK Version: \(self.factory?.version ?? "unknown") "
         
         guard let client = self.client else {
             return
