+++ Example/Split/GetTreatmentsViewController.swift
@@ -21,7 +21,7 @@ class GetTreatmentsViewController: UIViewController {
     @IBOutlet weak var evaluateActivityIndicator: UIActivityIndicatorView!
     
     var factory: SplitFactory?
-    var client: SplitClientProtocol?
+    var client: SplitClient?
 
     @IBAction func evaluate(_ sender: Any) {
         evaluate()
@@ -61,13 +61,14 @@ class GetTreatmentsViewController: UIViewController {
         let key: Key = Key(matchingKey: matchingKeyText, bucketingKey: bucketing)
         
         //Split Factory
-        self.factory = SplitFactory(apiKey: authorizationKey, key: key, config: config)
+        let builder = DefaultSplitFactoryBuilder()
+        self.factory = builder.setApiKey(authorizationKey).setKey(key).setConfig(config).build()
         
         //Split Client
-        self.client = self.factory?.client()
+        self.client = self.factory?.client
         
         //Showing sdk version in UI
-        self.sdkVersion?.text = "SDK Version: \(self.factory?.version() ?? "unknown") "
+        self.sdkVersion?.text = "SDK Version: \(self.factory?.version ?? "unknown") "
         
         guard let client = self.client else {
             return
