+++ Split/Api/SplitClientConfig.swift
@@ -144,169 +144,3 @@ public class SplitClientConfig: NSObject {
      */
     @objc public var impressionListener: SplitImpressionListener?
 }
-
-// MARK: Deprecated methods
-/**
- All this methods are deprecated and will be removed on next versions. Should use properties instead.
- */
-extension SplitClientConfig {
-    @available(*, deprecated, renamed: "sdkReadyTimeOut", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead.")
-    public func readyTimeOut(_ readyInMillis: Int){
-        self.sdkReadyTimeOut = readyInMillis
-    }
-
-    @available(*, deprecated, renamed: "sdkReadyTimeOut", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead")
-    public func getReadyTimeOut() -> Int {
-        return self.sdkReadyTimeOut
-    }
-
-    @available(*, deprecated, renamed: "featuresRefreshRate", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead")
-    public func featuresRefreshRate(_ rr: Int) {
-        self.featuresRefreshRate = rr
-    }
-
-    @available(*, deprecated, renamed: "featuresRefreshRate", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead")
-    public func getFeaturesRefreshRate() -> Int {
-        return self.featuresRefreshRate
-    }
-
-    @available(*, deprecated, renamed: "segmentsRefreshRate", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead")
-    public func segmentsRefreshRate(_ rr: Int) {
-        self.segmentsRefreshRate = rr
-    }
-
-    @available(*, deprecated, renamed: "segmentsRefreshRate", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead")
-    public func getSegmentsRefreshRate() -> Int {
-        return self.segmentsRefreshRate
-    }
-
-    @available(*, deprecated, renamed: "impressionRefreshRate", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead")
-    public func impressionRefreshRate(_ rr: Int){
-        self.impressionRefreshRate = rr
-    }
-
-    @available(*, deprecated, renamed: "impressionRefreshRate", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead")
-    public func getImpressionRefreshRate() -> Int {
-        return self.impressionRefreshRate
-    }
-
-    @available(*, deprecated, renamed: "impressionsChunkSize", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead")
-    public func impressionsChunkSize(_ cs: Int64) {
-        self.impressionsChunkSize = cs
-    }
-
-    @available(*, deprecated, renamed: "impressionsChunkSize", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead")
-    public func getImpressionsChunkSize() -> Int64 {
-        return self.impressionsChunkSize
-    }
-
-    @available(*, deprecated, renamed: "impressionsQueueSize", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead")
-    public func impressionsQueueSize(_ qs: Int) {
-        self.impressionsQueueSize = qs
-    }
-
-    @available(*, deprecated, renamed: "impressionsQueueSize", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead")
-    public func getImpressionsQueueSize() -> Int {
-        return self.impressionsQueueSize
-    }
-
-    @available(*, deprecated, renamed: "connectionTimeout", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead")
-    public func connectionTimeout(_ to: Int) {
-        self.connectionTimeout = to
-    }
-
-    @available(*, deprecated, renamed: "connectionTimeout", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead")
-    public func getConnectionTimeout() -> Int {
-        return self.connectionTimeout
-    }
-
-    @available(*, deprecated, renamed: "apiKey", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead")
-    public func apiKey(_ k: String) {
-        SecureDataStore.shared.setToken(token: k)
-    }
-
-    @available(*, deprecated, renamed: "apiKey", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead")
-    public func getApiKey() -> String {
-        return self.apiKey
-    }
-
-    @available(*, deprecated, renamed: "trafficType", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead")
-    public func trafficType(_ tt: String){
-        self.trafficType = tt
-    }
-
-    @available(*, deprecated, renamed: "trafficType", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead")
-    public func getTrafficType() -> String? {
-        return self.trafficType
-    }
-
-    @available(*, deprecated, renamed: "eventsFirstPushWindow", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead")
-    public func eventsFirstPushWindow(_ pw: Int){
-        self.eventsFirstPushWindow = pw
-    }
-
-    @available(*, deprecated, renamed: "eventsFirstPushWindow", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead")
-    public func getEventsFirstPushWindow() -> Int {
-        return self.eventsFirstPushWindow
-    }
-
-    @available(*, deprecated, renamed: "eventsPushRate", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead")
-    public func eventsPushRate(_ pr: Int){
-        self.eventsPushRate = pr
-    }
-
-    @available(*, deprecated, renamed: "eventsPushRate", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead")
-    public func getEventsPushRate() -> Int {
-        return self.eventsPushRate
-    }
-
-    @available(*, deprecated, renamed: "eventsQueueSize", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead")
-    public func eventsQueueSize(_ qs: Int64) {
-        self.eventsQueueSize = qs
-    }
-
-    @available(*, deprecated, renamed: "eventsQueueSize", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead")
-    public func getEventsQueueSize() -> Int64 {
-        return self.eventsQueueSize
-    }
-
-    @available(*, deprecated, renamed: "eventsPerPush", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead")
-    public func eventsPerPush(_ pp: Int) {
-        self.eventsPerPush = pp
-    }
-
-    @available(*, deprecated, renamed: "eventsPerPush", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead")
-    public func getEventsPerPush() -> Int {
-        return self.eventsPerPush
-    }
-
-    @available(*, deprecated, renamed: "isDebugModeEnabled", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead.")
-    public func debug(_ d:Bool){
-        Logger.shared.isDebugModeEnabled = d
-    }
-
-    @available(*, deprecated, renamed: "isVerboseModeEnabled", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead.")
-    public func verbose(_ v:Bool){
-        Logger.shared.isVerboseModeEnabled = v
-    }
-
-    @available(*, deprecated, renamed: "sdkEndpoint", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead.")
-    public func sdkEndpoint(_ u: String) {
-        EnvironmentTargetManager.shared.sdkEndpoint = u
-    }
-
-    @available(*, deprecated, renamed: "eventsEndpoint", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead.")
-    public func eventsEndpoint(_ u: String) {
-        EnvironmentTargetManager.shared.eventsEndpoint = u
-    }
-    
-    @available(*, deprecated, renamed: "impressionListener", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead.")
-    public func setImpressionListener(_ il: @escaping SplitImpressionListener){
-        impressionListener = il
-    }
-    
-    @available(*, deprecated, renamed: "impressionListener", message: "This function was deprecated and will be removed in future versions. Please use equivalent propery instead.")
-    public func getImpressionListener() -> SplitImpressionListener? {
-        return impressionListener
-    }
-}
