--- Split/Api/SplitFactoryProtocol.swift
@@ -1,18 +1,0 @@
-//
-//  SplitFactoryProtocol.swift
-//  Pods
-//
-//  Created by Brian Sztamfater on 27/9/17.
-//
-//
-
-import Foundation
-
-@objc public protocol SplitFactoryProtocol {
-    
-    func client() -> SplitClientProtocol
-    
-    func manager() -> SplitManagerProtocol
- 
-    func version() -> String
-}
