+++ Split/Matchers/DependencyMatcher.swift
@@ -15,7 +15,7 @@ public class DependencyMatcher: BaseMatcher, MatcherProtocol  {
     
     var dependencyData: DependencyMatcherData?
     //--------------------------------------------------------------------------------------------------
-    public init(splitClient: SplitClient? = nil, negate: Bool? = nil, attribute: String? = nil , type: MatcherType? = nil, dependencyData: DependencyMatcherData?) {
+    public init(splitClient: DefaultSplitClient? = nil, negate: Bool? = nil, attribute: String? = nil , type: MatcherType? = nil, dependencyData: DependencyMatcherData?) {
         
         super.init(splitClient: splitClient, negate: negate, attribute: attribute, type: type)
         self.dependencyData = dependencyData
