+++ Split/Matchers/InSegmentMatcher.swift
@@ -14,7 +14,7 @@ public class InSegmentMatcher: BaseMatcher, MatcherProtocol {
     var data: UserDefinedSegmentMatcherData?
     
     //--------------------------------------------------------------------------------------------------
-    public init(data: UserDefinedSegmentMatcherData?, splitClient: SplitClient? = nil, negate: Bool? = nil, attribute: String? = nil , type: MatcherType? = nil) {
+    public init(data: UserDefinedSegmentMatcherData?, splitClient: DefaultSplitClient? = nil, negate: Bool? = nil, attribute: String? = nil , type: MatcherType? = nil) {
         
         super.init(splitClient: splitClient, negate: negate, attribute: attribute, type: type)
         self.data = data
