+++ Split/Matchers/PartOfSetMatcher.swift
@@ -12,7 +12,7 @@ public class PartOfSetMatcher: BaseMatcher, MatcherProtocol {
     var data: Set<String>?
     
     //--------------------------------------------------------------------------------------------------
-    public init(data:[String]?, splitClient: SplitClient? = nil, negate: Bool? = nil, attribute: String? = nil , type: MatcherType? = nil) {
+    public init(data:[String]?, splitClient: DefaultSplitClient? = nil, negate: Bool? = nil, attribute: String? = nil , type: MatcherType? = nil) {
         
         super.init(splitClient: splitClient, negate: negate, attribute: attribute, type: type)
